Add timeout support via embassy-time (#32)
This change adds functionality to timeout all i2c transactions in case
the fuel gauge gets stuck/takes too long to respond.

Default timeout is 100ms but it can be changed by passing in a timeout
value at driver construction time.

Includes the changes for retries and backoff delay too.

Uprev to v0.3.0 as this is a breaking change.

**NOTE:** This change includes a dependency on the _crates.io version of
embassy time._ This is done because crates in crates.io must be
semantically versioned, and not use git references. If you are using the
git ref version in your project's `Cargo.toml`, you will likely run into
multiple dependency issues. To resolve this, a new branch with the git
version of embassy-time will be maintained alongside the `main` branch,
please point your `Cargo.toml` to use that version instead.