Add PEC support to all reads and writes (#44)
The SMBus spec defines an optional feature known as Packet Error
Checking (PEC), which is simply a CRC-8 checksum over the entire
message, appended to the end of each message. This protects against data
corruption.

The fuel gauge can sometimes time itself out when it takes too long to
think of a response to a request, or the host may not ACK a fuel gauge
response quick enough, which will lead to the fuel gauge not responding
with the entire message. The PEC ensures that the entire message is sent
without corruptions or failure.

Closes #34 

EDIT: Notes from integration testing

- 1 byte registers defined in the datasheet are actually 2 bytes, and
  will only return PEC if you read 2 bytes
- mfg_info commands (and dataflash) return 32 bytes as far as the fuel
gauge is
  concerned, so in order to read the PEC byte, the full 32 bytes must be
  read, even if we just want the first 4 bytes
- Discovered critical dataflash bug where dataflash reads would be off
  by 1 byte and return the wrong data. Bug fixed here.