Add DataTooLarge error handling for public methods that take in a slice of bytes (#47)
The manufacture info register accessors are passed an arbitrarily sized
number of bytes as an argument. Enforce a limit on these methods in case
a slice that is too large is passed in.

Additionally, fix regression introduced with PEC PR with data staging
arrays not being large enough and overrunning the buffer.

Semver wants a breaking change because of new Error enum introduced, so
uprev to 0.7.0